+++ packages/lib/src/recommendations/tips/index.ts
@@ -1,0 +1,1 @@
+export * from "./energy-tips";
